requirements.txt
fastapi
uvicorn[standard]
websockets
sqlalchemy
alembic
authlib
python-jose[cryptography]
pydantic-settings
psycopg2-binary
python-multipart
starlette
itsdangerous
httpx
langchain
langchain-google-genai
chromadb
pyjwt
composio
composio_langchain